+++ src/Navigator.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createBottomTabNavigator, BottomTabBar, createAppContainer } from 'react-navigation'
 
-import HomeScreens from './components/HomeScreens'
+import OngoingEventScreens from './components/OngoingEventScreens'
 import EventScreens from './components/EventScreens'
 import ProfileMainScreen from './components/ProfileScreens/ProfileMainScreen'
 
@@ -20,7 +20,7 @@ const style = {
 
 const MainTabNavigator = createBottomTabNavigator({
   OngoingEvent : {
-    screen: HomeScreens,
+    screen: OngoingEventScreens,
     navigationOptions: {
       tabBarLabel: 'SUKELLUS',
       tabBarOptions: style,
